separate fill() to sub-functions functionally
